requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attributedict==0.3.0
attrs==23.1.0
bitarray==2.8.2
blessings==1.7
cachetools==5.4.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
coti-contracts-examples==0.1.0
coti-sdk==1.0.2
coverage==7.6.1
cryptography==3.4.8
cytoolz==0.12.2
deepdiff==7.0.1
distlib==0.3.8
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
filelock==3.15.4
frozenlist==1.4.0
hexbytes==0.3.1
humanfriendly==10.0
idna==3.4
inspecta==0.1.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lru-dict==1.3.0
multidict==6.0.4
ordered-set==4.1.0
packaging==24.1
parsimonious==0.9.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==4.25.0
py-solc-x==1.1.1
pycparser==2.22
pycryptodome==3.19.0
Pygments==2.18.0
pyproject-api==1.7.1
python-dotenv==1.0.1
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rootpath==0.1.1
rpds-py==0.12.0
semantic-version==2.10.0
six==1.16.0
termcolor==2.4.0
tomli==2.0.1
toolz==0.12.0
tox==4.18.0
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.26.3
web3==6.11.2
websockets==12.0
yarl==1.9.2
